Implement Find/Replace feature
Not quite polished but most of the mechanics in place. User can enter
a term then see a view with all content items that match the term. From
this view the user can enter the replacement term and also exclude any
content items. Once ready, they can perform the find and replace, which
once complete will show a success message.